CXP-1105: move webhook commands
