+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/cli.yml
@@ -12,10 +12,3 @@ services:
             - '@akeneo_connectivity_connection.persistence.query.purge_connection_errors'
         tags:
             - {name: 'console.command'}
-
-    Akeneo\Connectivity\Connection\Infrastructure\Cli\PurgeEventsApiLogsCommand:
-        arguments:
-            - '@Akeneo\Connectivity\Connection\Infrastructure\Webhook\EventsApiDebug\Persistence\PurgeEventsApiSuccessLogsQuery'
-            - '@Akeneo\Connectivity\Connection\Infrastructure\Webhook\EventsApiDebug\Persistence\PurgeEventsApiErrorLogsQuery'
-        tags:
-            - {name: 'console.command'}
